+++ app/Services/Cart/Storage/Session.php
@@ -5,6 +5,10 @@ namespace App\Services\Cart\Storage;
 
 use Illuminate\Session\Store;
 
+/**
+ * Class Session
+ * Implements the logic of storing the shopping cart data in a session.
+ */
 class Session implements Storage
 {
     /**
@@ -12,6 +16,11 @@ class Session implements Storage
      */
     private $session;
 
+    /**
+     * Storage key.
+     *
+     * @var string
+     */
     private $key = 'cart';
 
     public function __construct(Store $session)
@@ -19,7 +28,10 @@ class Session implements Storage
         $this->session = $session;
     }
 
-    public function put(int $serverId, int $productId, float $amount): void
+    /**
+     * {@inheritdoc}
+     */
+    public function put(int $serverId, int $productId, int $amount): void
     {
         $server = $this->retrieveServer($serverId);
 
@@ -31,26 +43,41 @@ class Session implements Storage
         $this->session->put("{$this->key}.{$serverId}", $server);
     }
 
-    public function retrieve(int $serverId, int $productId): ?float
+    /**
+     * {@inheritdoc}
+     */
+    public function retrieve(int $serverId, int $productId): ?int
     {
         return $this->session->get("{$this->key}.{$serverId}.{$productId}");
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function retrieveServer(int $serverId): ?array
     {
         return $this->session->get("{$this->key}.{$serverId}");
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function remove(int $serverId, int $productId): bool
     {
         return (bool)$this->session->remove("{$this->key}.{$serverId}.{$productId}");
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function removeServer(int $serverId): bool
     {
         return (bool)$this->session->remove("{$this->key}.{$serverId}");
     }
 
+    /**
+     * {@inheritdoc}
+     */
     public function getKey(): string
     {
         return $this->key;
